requirements.txt
requests==2.32.3
python-telegram-bot==21.7
python-dotenv==1.0.1
anyio==4.6.2.post1